dicom-nifti conversion functional
